requirements.txt
file:///Users/shurrey/local/langchain/libs/core
file:///Users/shurrey/local/langchain/libs/community
file:///Users/shurrey/local/langchain/libs/text-splitters
file:///Users/shurrey/local/langchain/libs/langchain
file:///Users/shurrey/local/langchain/libs/partners/box
box-sdk-gen[JWT]
openai
chromadb
langchain-openai
tiktoken
python-dotenv
bs4
langchainhub
langgraph
jupyterlab
notebook
voila